Split wiki out of task_handler.py
